DG added api call for searched users in api file
